requirements.txt
fastapi>=0.111.0
uvicorn==0.30.0
python-multipart>=0.0.6
requests>=2.31.0
streamlit>=1.28.0
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.20.0
transformers>=4.40.0
Pillow>=10.0.0
loguru>=0.7.0
